Move copy logic to copy.rs
